update ipddp docs
